backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
drf-yasg==1.21.7
Pillow==10.2.0
django-filter==23.5
django-allauth==0.61.1
dj-rest-auth==5.0.2
stripe==7.11.0
lemon-squeezy==0.1.0 